Make compact 64bit compatible. Still need to test if this is valid.

+++ cudpp/include/cudpp.h
@@ -475,7 +475,7 @@ CUDPPResult cudppSegmentedScan(CUDPPHandle        planHandle,
 CUDPP_DLL
 CUDPPResult cudppCompact(CUDPPHandle        planHandle,
                          void               *d_out, 
-                         size_t             *d_numValidElements,
+                         unsigned int       *d_numValidElements,
                          const void         *d_in, 
                          const unsigned int *d_isValid,
                          size_t             numElements,
